pyproject.toml
[build-system]
requires = ["setuptools>=77.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "glyph-agents"
version = "0.2.0"
description = "Vendor Agnostic Agent SDK"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Louis Tricot", email = "louistricotkarl@gmail.com" },
]
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
]
license = "Apache-2.0"
license-files = ["LICEN[CS]E*"]
dependencies = [
    "claude-agent-sdk",
    "openai-agents",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.24,<1",
    "ruff>=0.14.10,<0.15",
]

[project.scripts]
glyph = "glyph.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
exclude = [
    ".venv*",
    ".git",
    "__pycache__",
    ".eggs",
    "*.egg",
    "build",
]
lint.ignore = [
    "A003",
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D200",
    "D202",
    "D204",
    "D205",
    "D208",
    "D210",
    "D300",
    "D400",
    "D401",
    "D403",
    "D404",
    "D413",
    "E203",
    "E231",
    "E721",
    "G201",
]
line-length = 120
lint.preview = true
lint.explicit-preview-rules = true
lint.select = [
    "A",
    "D",
    "E",
    "F",
    "G",
    "I",
    "PLW0244",
    "UP006",
    "W",
]
lint.unfixable = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.ruff.lint.isort]
force-single-line = true
lines-after-imports = 2
force-sort-within-sections = true
known-first-party = ["glyph"]
relative-imports-order = "furthest-to-closest"
